test(stepped-tracker): move unit tests to __tests__ folder
